% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inbox.R
\name{inbox}
\alias{inbox}
\title{Defining Subset lying in a Box}
\usage{
inbox(X, fixbox_metric = NULL, fixbox_nom = NULL, fixbox_na = NULL)
}
\arguments{
\item{X}{\code{\link{data.frame}} from which the subset should be taken.}

\item{fixbox_metric}{one row of a \code{box_metric} from a "\code{peel}"-object.}

\item{fixbox_nom}{one element of a \code{box_nom} list from a "\code{peel}"-object.}

\item{fixbox_na}{one element of a \code{box_na} list from a "\code{peel}"-object.}
}
\value{
\code{inbox} returns a logical vector indicating the observations that lie in the box.
}
\description{
Help function that returns the subset of the whole data set which is contained in one box (\code{fixbox}). This function is needed in other functions like \code{\link{PRIM_peel_bs}}.
}
\details{
The arguments \code{fixbox_metric}, \code{fixbox_nom}, \code{fixbox_na} can also come from a "\code{fixbox}"-object.
}

