% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inter_diss.R
\name{inter_diss}
\alias{inter_diss}
\title{Interbox Dissimilarity}
\usage{
inter_diss(fixbox1, fixbox2, data)
}
\arguments{
\item{fixbox1}{first object of class "\code{fixbox}".}

\item{fixbox2}{second object of class "\code{fixbox}".}

\item{data}{original data used to calculate the supports needed for the interbox dissimilarity. If this argument is missing \code{data_orig} from \code{fixbox1} is used as \code{data}.}
}
\description{
This Function calculates the interbox dissimilarity between two boxes defined on one data set.
}
\details{
The interbox dissimilarity is a diagnostic tool of PRIM measuring the dissimilarity between two boxes \eqn{B_k} and \eqn{B_l}.
It is defined as the difference between the smallest box \eqn{B_{kl}} that covers both boxes and the support of their union.

The interbox dissimilarity can assume values between 0 and 1.
While nested boxes have a dissimilarity of 0, it gets bigger the more different the two boxes are.
}

