% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.peel.R
\name{plot.peel}
\alias{plot.peel}
\title{Plotting a Trajectory}
\usage{
\method{plot}{peel}(x, ...)
}
\arguments{
\item{x}{object of class "\code{peel}".}

\item{...}{further arguments of the \code{\link{plot}}-function.}
}
\description{
This funcion plots the trajectory of a "\code{peel}"-object. In the trajectory the target functions evaluated on the data in the box are plotted against the supports beta.
}

