\name{prim-internal}

\alias{in.box}
\alias{counts.box}
\alias{prim.which.box}

\title{Internal functions in the prim library}
\description{
These functions are user-level but which the user is not required to use directly.
}

\value{The user is not required to use directly these outputs.}

\keyword{internal}
