% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRIM_peel.R
\name{PRIM_peel}
\alias{PRIM_peel}
\title{Peeling-Function}
\usage{
PRIM_peel(formula, data, peel_alpha = 0.05, beta_min = 0.01,
  target = mean, alter_crit = TRUE, use_NAs = TRUE)
}
\arguments{
\item{formula}{an object of class "\code{\link{formula}}" with a response but no interaction terms.
It indicates the response over which the target function should be maximized and the covariates that are used for the later box definitions.
If this argument is missing, the argument data is used as the \code{\link{model.frame}}.}

\item{data}{an object of class \code{\link{data.frame}} containing the variables named in the formula.}

\item{peel_alpha}{alpha-fraction used for the peeling.}

\item{beta_min}{minimum support that one Box should have (stop-criterion).}

\item{target}{target-function to be maximized.}

\item{alter_crit}{logical. If \code{TRUE} the alternative criterion is used for peeling.}

\item{use_NAs}{logical. If \code{TRUE} observations with missing values are included in the analysis.}
}
\value{
\code{PRIM_peel} returns an object of class "\code{peel}", which is a list containing at least the following components:

\item{f}{vector of the target functions evaluated on the box at each peeling step.}

\item{beta}{vector of the supports beta of the boxes at each peeling step.}

\item{box}{a \code{\link{data.frame}} defining the borders of the boxes. Each row belongs to one peeling step. The columns with "\code{min.}" and "\code{max.}" describe the lower and upper boundaries of the at least ordinal covariates. Therefore the value taken is the last one that is \bold{not} included in the current box.

For the nominal variables there are columns for every category they can take. If the category is removed from the box the value \code{FALSE} is taken. The names of these columns are structured like: \code{<variable name>.<category>}

For each variable with missing values (only if \code{use_NAs = TRUE}) there is also a column taking the value \code{FALSE} if the \code{NA}s of this variable are removed from the current box. The names of these columns are structured like: \code{<variable name>.NA}
}

\item{box_metric, box_nom, box_na}{easier to handle definitions of the boxes for other functions}

\item{subsets}{\code{list} of logical vectors indicating the subsets at each peeling step (i.e. the observations that lie in the box)}

\item{data_orig}{original dataset that is used for the peeling.}
}
\description{
This function is an implementation of the (singular) Peeling-Algorithm as suggested by Friedman and Fisher (1999). In each iteration the fraction alpha is peeled from one edge of the current box.
}
\details{
The outcome of the \code{formula} can either be numeric, logical or a survival object (see \code{\link{Surv}}). If it is a survival object the \code{target} is set to the number of events per amount of time.

This function is the main part of the multiple Version \code{PRIM_peel_bs}.
}
\examples{
# generating random data:
set.seed(123)
n <- 500
x1 <- runif(n = n, min = -1)
x2 <- runif(n = n, min = -1)
x3 <- runif(n = n, min = -1)
cat <- as.factor(sample(c("a","b","c", "d"), size = n, replace = TRUE))
wsk <- (1-sqrt(x1^2+x2^2)/sqrt(2))
y <- as.logical(rbinom(n = n, prob = wsk, size = 1))
dat <- cbind.data.frame(y, x1, x2, x3, cat)
#plot(dat$x1, dat$x2, col=dat$y+1, pch=16)
remove(x1, x2, x3, y, wsk, cat, n)

# apply the PRIM_peel function:
prim <- PRIM_peel(formula=y ~ ., data=dat, beta_min = .01, peel_alpha = .1)
plot(prim) # trajectory
prim$box # box definitions

}
\references{
Friedman, J. H. and Fisher, N. I., 'Bump hunting in high-dimensional data', Statistics and Computing 9 (2) (1999), 123-143

Ott, A. and Hapfelmeier, A., 'Nonparametric Subgroup Identification by PRIM and CART: A Simulation and Application Study', Computational and Mathematical Methods in Medicine, vol. 2017 (2017), 17 pages, Article ID 5271091
}

