% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.PRIM.R
\name{print.PRIM}
\alias{print.PRIM}
\title{Printing a "PRIM"-Object}
\usage{
\method{print}{PRIM}(x, ...)
}
\arguments{
\item{x}{object of class "\code{PRIM}".}

\item{...}{further arguments of the \code{\link{print}}-function.}
}
\description{
This function prints a \code{PRIM}-object. It is a method for the generic function \code{print} of class "\code{PRIM}".
}

