% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relfreq.R
\name{relfreq}
\alias{relfreq}
\title{Relative Frequency Plots}
\usage{
relfreq(fixbox, plot_together = TRUE, ...)
}
\arguments{
\item{fixbox}{object of class "\code{fixbox}" to be used to create the relative frequency plots.}

\item{plot_together}{logical. If \code{TRUE} all plots are drawn together in one figure by \code{par(mfrow = ...)}.}

\item{...}{further arguments of the functions \code{\link{plot}} and \code{\link{barplot}}.}
}
\description{
This Function creates the relative frequency plots of a box defined by \code{\link{define_fixbox}}.
}
\details{
Relative frequency plots are a diagnostic tool for a result of PRIM (\code{fixbox}).
These are histograms or barplots illustrating ratios of distributions \eqn{r_{jk}} for each variable \eqn{x_j}.
This ratio is defined by the distribution of \eqn{x_j} in one box \eqn{B_k} divided by the distribution of this variable over the whole data set.
}
\examples{
# generating random data:
set.seed(123)
n <- 500
x1 <- runif(n = n, min = -1)
x2 <- runif(n = n, min = -1)
x3 <- runif(n = n, min = -1)
cat <- as.factor(sample(c("a","b","c", "d"), size = n, replace = TRUE))
wsk <- (1-sqrt(x1^2+x2^2)/sqrt(2))
y <- as.logical(rbinom(n = n, prob = wsk, size = 1))
dat <- cbind.data.frame(y, x1, x2, x3, cat)
plot(dat$x1, dat$x2, col=dat$y+1, pch=16)
remove(x1, x2, x3, y, wsk, cat, n)

# apply the PRIM function to find the best box with a support of at least 0.1:
p <- PRIM(y~., data=dat, beta_min = 0.1, max_boxes = 1, print_position = FALSE)

# relative frequency plots:
relfreq(p$fixboxes[[1]])

}
\seealso{
\code{\link{define_fixbox}}, \code{\link{PRIM}}
}

