% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_dominated.R
\name{remove_dominated}
\alias{remove_dominated}
\title{Remove Dominated Boxes from a Peeling Output}
\usage{
remove_dominated(prim, sort = TRUE, dup.rm = TRUE)
}
\arguments{
\item{prim}{object of class "\code{peel}".}

\item{sort}{logical. If \code{TRUE} the boxes in the output are sorted decreasing by their box supports.}

\item{dup.rm}{logical. Tf \code{TRUE} duplicated boxes in the output are removed. Boxes are duplicated if another box has exactly the same target and support.}
}
\value{
This function returns another "\code{peel}"-object having the same structure as the input object, just without the dominated boxes.
}
\description{
This function removes all boxes from one "\code{peel}"-object that are dominated by another box.
I.e. there is at least one other box with a better output and a bigger support beta.
}
\details{
Dominated boxes mainly occur in outputs of the multiple peeling function (\code{\link{PRIM_peel_bs}}).
So this function is practically only useful for "\code{peel}"-objects beeing the results of such a function.
Without the dominated boxes the object gets much smaller by keeping all the relevant boxes.
Also the trajectory (see \code{\link{plot.peel}}) without the not useful dominated boxes looks much clearer.
}
\seealso{
\code{\link{PRIM_peel_bs}}, \code{\link{plot.peel}}
}

