% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_select.R
\name{var_select}
\alias{var_select}
\title{Sequential Relevance of Box-defining Variables}
\usage{
var_select(fixbox)
}
\arguments{
\item{fixbox}{object of class "\code{fixbox}" to be used to create the table of the sequential variable relevances.}
}
\value{
\code{var_select} returns a list with the following components:
\item{tab}{table showing the sequential relevances of the variables used in the definition of the \code{fixbox}.}
\item{fixboxes}{\code{list} of the boxes defined at every iteration step (by removing of one variable).}
}
\description{
This function creates a table of the sequential relevance of the variables used in the definition of a box defined by \code{\link{define_fixbox}}.
}
\details{
This function does a sequential removal off all variables defining one box.
In each iteration every variable left in the box definition is tried to be completley left out of the definition.
From all variables this one is removed, which causes the least decrease of the target function evaluated on all observations lying in the box.
}
\examples{
# generating random data:
set.seed(123)
n <- 500
x1 <- runif(n = n, min = -1)
x2 <- runif(n = n, min = -1)
x3 <- runif(n = n, min = -1)
cat <- as.factor(sample(c("a","b","c", "d"), size = n, replace = TRUE))
wsk <- (1-sqrt(x1^2+x2^2)/sqrt(2))
y <- as.logical(rbinom(n = n, prob = wsk, size = 1))
dat <- cbind.data.frame(y, x1, x2, x3, cat)
plot(dat$x1, dat$x2, col=dat$y+1, pch=16)
remove(x1, x2, x3, y, wsk, cat, n)

# apply the PRIM function to find the best box with a support of at least 0.1:
p <- PRIM(y~., data=dat, beta_min = 0.1, max_boxes = 1, print_position = FALSE)

# sequential variable relevances:
var_select(p$fixboxes[[1]])$tab

}
\seealso{
\code{\link{define_fixbox}}, \code{\link{PRIM}}
}

