\name{prim-internal}

\alias{dmvnorm.mixt.dd}
\alias{in.box}
\alias{counts.box}
\alias{prim.thresh.mixt.dd} 
\alias{prim.which.box}
\alias{rmvnorm.mixt.dd}
\alias{thresh.mixt.dd}

\title{Internal functions in the prim library}
\description{
These functions are user-level but which the user is not required to use directly.
}


\keyword{internal}
