\name{quasiflow}

\docType{data}

\alias{quasiflow}

\title{Quasi flow cytometry data}

\description{
  This data set is simulated data from two normal mixture
  distrbutions, mimicking a flow cytometry
  data set. It contains 1000 observations from an HIV+ patient and
  1000 observations an HIV- patient.    
}

\usage{data(quasiflow)}

\format{\code{quasflow} is a matrix with 6 columns and 2000 rows.
  Each row corresponds to measurements for one cell.
  The first 5 columns are flow cytometric measurements and the sixth
  column is a binary indicator, with 1 = HIV+ and -1 = HIV-.
}

\source{
Randomly generated by package author.  
}


\keyword{datasets}

